+++ melhor-consulta-spring/src/main/java/com/soaresdev/melhorconsultaspring/repository/EspecialidadeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EspecialidadeRepository extends JpaRepository<Especialidade, Long> {
+public interface EspecialidadeRepository extends JpaRepository<Especialidade, Integer> {
 }
